chore(installer): extract shared test harness to laravel-swarm-installer-testkit
The swarm:install* test harness (InstallerTestCase, InstallerRunResult,
DoubleRunResult) lived in tests/Installer/ (autoload-dev, package-private
in Composer), so the Pulse companion package had to duplicate ~250 lines
of it. Extracted into a standalone require-dev-only package instead —
orchestra/testbench stays out of core's production require, and every
companion (Pulse now, Filament/MCP later) can require-dev the kit
directly rather than copying it.

getPackageProviders() is no longer swarm-specific; SwarmInstallerTestCase
here supplies the swarm provider list on top of the shared base class.

Closes #355.
