+++ tests/Installer/InstallDurableCommandTest.php
@@ -3,10 +3,10 @@
 declare(strict_types=1);
 
 use BuiltByBerry\LaravelSwarm\Commands\Install\InstallDurableCommand;
-use BuiltByBerry\LaravelSwarm\Tests\Installer\InstallerTestCase;
+use BuiltByBerry\LaravelSwarm\Tests\Installer\SwarmInstallerTestCase;
 use Illuminate\Contracts\Config\Repository as ConfigRepository;
 
-uses(InstallerTestCase::class);
+uses(SwarmInstallerTestCase::class);
 
 /**
  * Feature tests for `swarm:install:durable`.
